+++ constants.js
@@ -1,0 +1,43 @@
+const Messages = {
+  "AncestorIncompatibility": {
+    "message": `Anchor and focus nodes of selection should have the same parent node
+    or the same grandparent if one of the nodes' parent is one of <b>, <ul>, <i>`
+  }
+} 
+
+const BlockElementNames = [
+  "ADDRESS",
+  "ARTICLE",
+  "ASIDE",
+  "BLOCKQUOTE",
+  "CANVAS",
+  "DD",
+  "DIV",
+  "DL",
+  "DT",
+  "FIELDSET",
+  "FIGCAPTION",
+  "FIGURE",
+  "FOOTER",
+  "FORM",
+  "H1",
+  "H2",
+  "H3",
+  "H4",
+  "H5",
+  "H6",
+  "HEADER",
+  "HR",
+  "LI",
+  "MAIN",
+  "NAV",
+  "NOSCRIPT",
+  "OL",
+  "P",
+  "PRE",
+  "SECTION",
+  "TABLE",
+  "TFOOT",
+  "UL",
+  "VIDEO"
+]